+++ bin/CommonConst.js
@@ -1,0 +1,17 @@
+/**
+* 共通定数
+**/
+var CommonConst = {
+
+	//ページID
+		//view
+	,	PAGE_ID_VIEW_CHAT				: 'chat'
+	
+	//テーブル名
+	//,	TABLE_NAME_QUESTION				: 'QUESTION'
+	//,	TABLE_NAME_EXAM					: 'EXAM'
+	//,	TABLE_NAME_USER					: 'USER'
+	//,	TABLE_NAME_QUESTION_CATEGORY	: 'QUESTION_CATEGORY'
+	//,	TABLE_NAME_STUDY_HISTORY		: 'STUDY_HISTORY'
+};
+module.exports = CommonConst;
